backend/requirements.txt
Flask
gunicorn
python-dotenv
openai
langdetect

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1